// src/interfaces/antecedentesprima-afp.model.ts
export class AntecedentesPrimaAfp {
    NumPoliza?: string;
    CodTippension?: string;
    NumIdenafi?: string;
    GlsTipoidencor?: string;
    CodCuspp?: string;
    CodTipren?: string; 
    NumMesdif?: number;
    CodModalidad?: string; 
    NumMesgar?: number;
    MtoBono?: number;
    MtoPriuni?: number;
    MtoPension?: number;
    MtoPensiongar?: number;
    MtoValprepentmp?: number;
    MtoRentatmpafp?: number;
    MtoCtaindafp?: number;
    MtoResmat?: number;
    MtoPenanual?: number;
    MtoRmpension?: number;
    MtoRmgtosep?: number;
    MtoRmgtoseprv?: number;
    MtoPriunisim?: number;
    PrcTasarprt?: number;
    GlsPension?: string;
    GlsRenta?: string;
    GlsModalidad?: string;
    GlsNomben?: string;
    GlsNomsegben?: string;
    GlsPatben?: string;
    GlsMatben?: string;
    CodMoneda?: string;
    FecAcepta?: string;
    FecDev?: string;
    MtoValmoneda?: number;
    MtoPriunidif?: number;
    PrcRentatmp?: number;
    MtoSumpension?: number;
    CodTipreajuste?: string;
    MtoValreajustetri?: number;
    MtoValreajustemen?: number;
    GlsTipreajuste?: string;
    CodMontipreaju?: string;
    GlsMontipreaju?: string;
    NumMesesc?: number;
    PrcRentaesc?: number;
    PrcPension?: number;
    MtoPrirecpesos?: number;
    MtoValmonedarec?: number;
    Prc_TasaRPRT?: number;
    PrcFacvar?: number;
    MtoPriafp?: number;
    MtoPensionafp?: number;
    MtoPricia?: number;
    MtoSumpensionafp?: number;
    NomCompleto?: string;
    TipoPensionCom?: string;
    TipoRentaCom?: string;
    TipoModalidadCom?: string;
    TipoMonedaCom?: string;
    MtoSumPensionDef?: number;
    FecTraspaso?: string;
    CodUsuariocrea?: string;
    MtoPenvarfon?: number;
    MtoPenvartc?: number;
    MtoPenvarfont?: number;
    FecVigencia?: string;
    CodLiquidacion?: string;
    CodFactura?: string;
    CodRenvit?: string;
    MtoSumpensioninf?: number;
    FecEmision?: string;
    NumCotizacion?: string;
    Mensaje?: string;
    Usuario?: string;
}

export class Moneda{
    SimboloMoneda?: string="";
    CodScomp?: string="";
}

export class Ubigeo{
    COD_UBIGEOCORRESP?: number=0;
    GLS_DIRECCIONCORRESP?: string="";
    NUM_POLIZA?: string="";
    Usuario?: string="";
    GlsCorreo?: string="";
    GlsCorreo2?: string="";
    GlsCorreo3?: string="";
    GlsFono?: string="";
    GlsFono2?: string="";
    GlsFono3?: string="";
    TipPension?: string="";
    Boleta?: string="";
    Proteccion?: string="";
    
}